- fixed more style cop stuff
