+++ Source/OxyPlot/Series/FinancialSeries/OhlcvItem.cs
@@ -9,6 +9,8 @@
 
 namespace OxyPlot.Series
 {
+    using System.Collections.Generic;
+
     /// <summary>
     /// Represents an item in a <see cref="CandleStickAndVolumeSeries" />.
     /// </summary>
@@ -19,7 +21,6 @@ namespace OxyPlot.Series
         /// </summary>
         public static readonly OhlcvItem Undefined = new OhlcvItem(double.NaN, double.NaN, double.NaN, double.NaN, double.NaN);
 
-
         /// <summary>
         /// Initializes a new instance of the <see cref="OhlcvItem" /> class.
         /// </summary>
@@ -27,15 +28,14 @@ namespace OxyPlot.Series
         { 
         }
 
-
         /// <summary>
         /// Initializes a new instance of the <see cref="OxyPlot.Series.OhlcvItem"/> class.
         /// </summary>
         /// <param name="x">The x coordinate / time.</param>
-        /// <param name="open">Open.</param>
-        /// <param name="high">High.</param>
-        /// <param name="low">Low.</param>
-        /// <param name="close">Close.</param>
+        /// <param name="open">Open value.</param>
+        /// <param name="high">High value.</param>
+        /// <param name="low">Low value.</param>
+        /// <param name="close">Close value.</param>
         /// <param name="buyvolume">Buy volume.</param>
         /// <param name="sellvolume">Sell volume.</param>
         public OhlcvItem(
@@ -64,13 +64,11 @@ namespace OxyPlot.Series
         /// <value>The X value.</value>
         public double X { get; set; }
 
-
         /// <summary>
         /// Gets or sets the open value.
         /// </summary>
         /// <value>The open value.</value>
         public double Open { get; set; }
-
 
         /// <summary>
         /// Gets or sets the high value.
@@ -78,13 +76,11 @@ namespace OxyPlot.Series
         /// <value>The high value.</value>
         public double High { get; set; }
 
-
         /// <summary>
         /// Gets or sets the low value.
         /// </summary>
         /// <value>The low value.</value>
         public double Low { get; set; }
-
 
         /// <summary>
         /// Gets or sets the close value.
@@ -92,20 +88,87 @@ namespace OxyPlot.Series
         /// <value>The close value.</value>
         public double Close { get; set; }
 
-
         /// <summary>
         /// Gets or sets the buy volume.
         /// </summary>
         public double BuyVolume { get; set; }
-
 
         /// <summary>
         /// Gets or sets the sell volume.
         /// </summary>
         public double SellVolume { get; set; }
 
-
         // Functions
+
+        /// <summary>
+        /// Find index of max(x) &lt;= target x in a list of OHLCV items
+        /// </summary>
+        /// <param name='items'>
+        /// vector of bars
+        /// </param>
+        /// <param name='targetX'>
+        /// target x.
+        /// </param>
+        /// <param name='guessIdx'>
+        /// initial guess.
+        /// </param>
+        /// <returns>
+        /// index of x with max(x) &lt;= target x or -1 if cannot find
+        /// </returns>
+        public static int FindIndex(List<OhlcvItem> items, double targetX, int guessIdx)
+        {
+            int lastguess = 0;
+            int start = 0;
+            int end = items.Count - 1;
+
+            while (start <= end)
+            {
+                if (guessIdx < start)
+                {
+                    return lastguess;
+                }
+                else if (guessIdx > end)
+                {
+                    return end;
+                }
+
+                var guessX = items[guessIdx].X;
+                if (guessX == targetX)
+                {
+                    return guessIdx;
+                }
+                else if (guessX > targetX)
+                {
+                    end = guessIdx - 1;
+                    if (end < start)
+                    {
+                        return lastguess;
+                    }
+                    else if (end == start)
+                    {
+                        return end;
+                    }
+                }
+                else
+                { 
+                    start = guessIdx + 1; 
+                    lastguess = guessIdx; 
+                }
+
+                if (start >= end)
+                {
+                    return lastguess;
+                }
+
+                var endX = items[end].X;
+                var startX = items[start].X;
+
+                var m = (double)(end - start + 1) / (endX - startX);
+                guessIdx = start + (int)((targetX - startX) * m);
+            }
+
+            return lastguess;
+        }
 
         /// <summary>
         /// Indicate whether is valid for rendering or not
